+++ src/components/main/CategoryButton.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function CategoryButton() {
+  return <div />;
+}
+
+export default CategoryButton;
